requirements.txt
pdfplumber==0.11.9
protobuf==7.34.0
Requests==2.32.5
streamlit==1.43.0
google-genai==0.1.0